requirements.txt
fastapi
uvicorn
request
tinydb